pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geminigrip"
version = "0.1.0"
description = "CLI to automate Google Gemini Deep Research using a persistent browser or OpenClaw."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.49.0",
]

[project.scripts]
geminigrip = "main:app"

[tool.setuptools.packages.find]
where = ["."]
packages = ["scripts"]
